Move LinterEntry to indexdigest.utils
