+++ indexdigest/utils.py
@@ -1,5 +1,5 @@
 """
-This module contains utility functions
+This module contains utility functions and classes
 """
 try:
     from urlparse import urlparse  # Python2
@@ -27,3 +27,24 @@ def parse_dsn(dsn):
         'passwd': parsed.password,
         'db': str(parsed.path).lstrip('/')
     }
+
+
+class LinterEntry(object):
+    """
+    Wraps a single linter entry. Various formatters may display this data differently.
+    """
+    def __init__(self, linter_type, table_name, message, context=None):
+        """
+        :type linter_type str
+        :type table_name str
+        :type message str
+        :type context dict
+        """
+        self.linter_type = linter_type
+        self.table_name = table_name
+        self.message = message
+        self.context = context
+
+    def __str__(self):
+        return '{table_name}: {message}'.format(
+            table_name=self.table_name, message=self.message)
